testing funcs to own file
