+++ files/food.txt
@@ -1,59 +1,30 @@
 I like to eat pizza, so I go to the pizzeria on Main Street.
-
 I like to eat hamburgers, so I go to the new burger joint downtown.
-
 I like to eat sushi, so I go to the Japanese restaurant near the park.
-
 I like to eat pasta, so I go to the Italian trattoria in town.
-
 I like to eat Chinese food, so I go to the authentic Chinese buffet in Chinatown.
-
 I like to eat tacos, so I go to the vibrant food truck near the park.
-
 I like to eat donuts, so I go to the cozy bakery downtown.
-
 I like to eat ramen, so I go to the university canteen where they serve the best ramen.
-
 I like to eat kebab, so I go to the Mediterranean grill with freshly grilled meats.
-
 I like to eat steak, so I go to the downtown steakhouse renowned for its ribeye.
-
 I like to eat croissants, so I go to the French patisserie known for its flaky pastries.
-
 I like to eat avocado toast, so I go to the campus café with a modern twist.
-
 I like to eat salad, so I go to the organic bistro that specializes in fresh, healthy dishes.
-
 I like to eat carbonara, so I go to the Italian trattoria that serves creamy pasta.
-
 I like to eat enchiladas, so I go to the small Mexican cantina with authentic flavors.
-
 I like to eat seafood chowder, so I go to the coastal bistro by the harbor.
-
 I like to eat gourmet tacos, so I go to the food truck that offers international twists.
-
 I like to eat fusion dishes, so I go to the art district café where different cuisines meet.
-
 I like to eat artisan pizza, so I go to the trendy pizza bar with creative toppings.
-
 I like to eat burgers, so I go to the retro diner with classic American flavors.
-
 I like to eat barbecue, so I go to the local smokehouse known for its slow-cooked ribs.
-
 I like to eat ice cream, so I go to the vintage ice cream parlor downtown.
-
 I like to eat falafel, so I go to the Middle Eastern restaurant with a lively atmosphere.
-
 I like to eat crepes, so I go to the charming French café that serves sweet and savory crepes.
-
 I like to eat curry, so I go to the Indian restaurant with a rich, aromatic menu.
-
 I like to eat pierogi, so I go to the Eastern European bistro with handmade specialties.
-
 I like to eat bagels, so I go to the cozy deli that serves them freshly baked every morning.
-
 I like to eat smoothies, so I go to the modern juice bar that offers a variety of blends.
-
 I like to eat steak frites, so I go to the bistro that specializes in French cuisine.
-
 I like to eat dim sum, so I go to the bustling Cantonese restaurant at lunchtime.
