+++ models/transformer.py
@@ -2,17 +2,7 @@ import math
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
-
-class RMSNorm(nn.Module):
-    """Root Mean Square Layer Normalization"""
-    def __init__(self, dim: int, eps: float = 1e-5):
-        super().__init__()
-        self.eps = eps
-        self.weight = nn.Parameter(torch.ones(dim))
-
-    def forward(self, x: torch.Tensor) -> torch.Tensor:
-        norm = x.pow(2).mean(dim=-1, keepdim=True).add(self.eps).sqrt()
-        return x / norm * self.weight
+from .utils import RMSNorm
 
 class MultiHeadAttention(nn.Module):
     """Standard Multi-Head Attention"""
