+++ models/utils.py
@@ -1,15 +1,16 @@
 import math
+import torch
 import torch.nn as nn
 
 # shared model utils, common layers
 
-
-
 class RMSNorm(nn.Module):
-    def __init__(self, dim, eps=1e-5):
+    """Root Mean Square Layer Normalization"""
+    def __init__(self, dim: int, eps: float = 1e-5):
         super().__init__()
         self.eps = eps
         self.weight = nn.Parameter(torch.ones(dim))
-    def forward(self, x):
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
         norm = x.pow(2).mean(dim=-1, keepdim=True).add(self.eps).sqrt()
-        return self.weight * (x / norm)
+        return (x / norm) * self.weight
